Update README.md
There was a class presentation in this course, in which we had to select a theme and then we had present it in clas as a group.My team members were me,Raj,Sanjeev and Karthik.We had picked up a topic from a book name computer organization by Hamacher.We found a topic from there name,Computer Peripherals.Going through this topic based presentation I had learnt about the different kinds of input devices and their functionings in daily life.And also I got to know how they work at basic level like for example data transmission from one device to another through exchange of electrical signals.The machines communicate through the electrical signals in their own language called bits.In that presentation, we discussed about input devices, output devices, common serial communication protocols,types of printers, graphic display devices,types of videos displays and so on.It was all good to work as a team in this course.Although we found difficult to find a topic for presenting in the class. But we had to find that kind of topic so that we be clear about that.Then we found this topic in our uploaded book on moodle.So,it was not easy to read and understand but after reading it twice and thrice,I got that.Then we made the presentation of it,learnt through the presentations and the questions asked in the class.